+++ index.js
@@ -2,7 +2,7 @@ const express = require('express')
 const exphbs = require('express-handlebars')
 const path = require('path')
 
-const homeRoutes = require('./routes/index')
+const homeRoutes = require('./routes/home')
 const flowersRoutes = require('./routes/flowers')
 const aboutRoutes = require('./routes/about')
 const contactsRoutes = require('./routes/contacts')
@@ -17,10 +17,11 @@ const hbs = exphbs.create({
 })
 
 app.engine('hbs', hbs.engine)
-app.set('view engine', hbs)
+app.set('view engine', 'hbs')
 app.set('views', 'views')
 
 app.use(express.static(path.join(__dirname, 'public')))
+
 app.use(express.urlencoded({extended: true}))
