+++ routes/home.js
@@ -1,9 +1,9 @@
 const {Router} = require('express')
 const router = Router()
-const path = require('path')
+// const path = require('path')
 
 router.get('/', (req, res) => {
-    res.sendFile(path.join(__dirname, '..', 'views', 'index.html'))
+    res.render('home')
 })
 
 module.exports = router
