+++ views/home.hbs
@@ -1,0 +1,1 @@
+<h1>main page</h1>
