--- views/index.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-<head>
-    <meta charset="UTF-8">
-    <title>Titled</title>
-    <link href="../style/style.css" rel="stylesheet" type="text/css" >
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300&display=swap" rel="stylesheet">
-</head>
-<body class="background-body">
-    <div class="background-header">
-        <header class="s1" >
-            <img src="../img/лого.png" class="img1" >
-                <nav class="nav">
-                    <ul>
-                        <a href="#"  >
-                            <li class="contacts">Контакты</li>
-                        </a>
-                        <a href="#"  >
-                            <li class="about ">О нас</li>
-                        </a>
-                        <a href="#"  >
-                            <li class="flowers  ">Цветы</li>
-                        </a>
-                        <a href="#"  >
-                            <li class="main ">Главная</li>
-                        </a>
-                    </ul>
-                </nav>
-                <div class="korzina">
-                    <img src="../img/shopping-cart%201.png" class="img2" >
-                    <a href="#" >
-                        <p>Корзина</p>
-                    </a>
-                </div>
-        </header>
-    </div>
-
-</body>
-</html>
