+++ views/layouts/main.hbs
@@ -1,0 +1,13 @@
+{{> head }}
+
+<body>
+
+{{> navbar}}
+
+<div class="background-body">
+    {{{ body }}}
+</div>
+
+
+{{> footer}}
+</body>
