+++ views/partials/navbar.hbs
@@ -1,0 +1,31 @@
+<div class="background-header">
+    <header class="s1" >
+        <img src="/img/logo.png" class="img1" alt="logo">
+        <nav class="nav">
+            <ul>
+                <li class="contacts">
+                    <a href="#"  >Контакты</a>
+                </li>
+
+                <li class="about ">
+                    <a href="#"  >О нас</a>
+                </li>
+
+
+                <li class="flowers  ">
+                    <a href="#"  >Цветы</a>
+                </li>
+
+                <li class="main ">
+                    <a href="#"  >Главная</a>
+                </li>
+            </ul>
+        </nav>
+        <div class="korzina">
+            <img src="/img/shopping-cart_1.png" class="img2" alt="logo"/>
+            <a href="#" >
+                <p>Корзина</p>
+            </a>
+        </div>
+    </header>
+</div>
